requirements.txt
click==7.1.2
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
redis==3.5.3
SQLAlchemy==1.3.22
Werkzeug==1.0.1
WTForms==2.3.3